Changed vertices type from uint16_t o uint32_t

+++ VulkanEngine/Buffers.cpp
@@ -290,7 +290,7 @@ std::vector<VertexBuffer::sVertex> VertexBuffer::vertices = {
 	{{-0.5f, 1.5f, -1.0f}, {1.0f, 1.0f, 1.0f}, {1.0f, 1.0f}, 0.0f}
 };
 
-const std::vector<uint16_t> VertexBuffer::indices = {
+const std::vector<uint32_t> VertexBuffer::indices = {
 	0, 1, 2, 2, 3, 0,
 	4, 5, 6, 6, 7, 4,
 	8, 9, 10, 10, 11, 8
